backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
strawberry-graphql[fastapi]
